build(docker): Add hack to cache rustup installation as a layer
This is because, due to the rust-toolchain.toml file, rustup would
automatically grab the latest Rust toolchain. The fact that this
happened during the cargo build phase, and therefore after
files were copied into the build image means that Docker would
not cache the toolchain, and it would be redownloaded every
time.
This could also be solved by using
mount=type=cache,target=/usr/local/rustup
on the RUN command for cargo build.